% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRepertoireEdgeDataframe.R
\name{createRepertoireEdgeDataframe}
\alias{createRepertoireEdgeDataframe}
\title{Create Immune Repertoire Network Edge Dataframe}
\usage{
createRepertoireEdgeDataframe(node_df, interaction.type = "hyperlink")
}
\arguments{
\item{node_df}{node dataframe outputted from createRepertoireNodeDataframe}

\item{interaction.type}{Column name containing the interaction type <default: 'hyperlink'>}
}
\value{
A dataframe with all pairwise edge interactions used to create a network.
}
\description{
Takes an input node dataframe of VDJ results (from createRepertoireNodeDataframe) and creates an edge dataframe which can be used to construct an Immune Repertoire network
}
\examples{
edge_df <- createRepertoireEdgeDataframe(node_df)
}
