% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRepertoireNodeDataframe.R
\name{createRepertoireNodeDataframe}
\alias{createRepertoireNodeDataframe}
\title{Create Immune Repertoire Network Node Dataframe}
\usage{
createRepertoireNodeDataframe(
  imm_results,
  patient.col = "patient",
  seqid.col = "unique_cell_id",
  locus.col = "locus",
  clone.col = "clone_id",
  cdr3.aa.col = "junction_aa",
  umi.count.col = "umi_count",
  read.count.col = "consensus_count",
  assay = "bcr",
  select.patients = NULL,
  expansion.thresh = 2
)
}
\arguments{
\item{imm_results}{results dataframe outputted from immcantation and merged using mergeVDJresults function}

\item{patient.col}{Column name containing the patient ID <default: 'patient'>}

\item{seqid.col}{Column name containing the sequence ID <default: 'unique_cell_id'>}

\item{locus.col}{Column name which contains the type of BCR or TCR chain <default: 'locus'>}

\item{clone.col}{Column name which contains the clone ID <default: 'clone_id'>}

\item{cdr3.aa.col}{Column name which contains the AA sequences of the junction including the CDR3 region <default: 'junction_aa'>}

\item{umi.count.col}{Column name which contains the UMI count per contig <default: 'umi_count'>}

\item{read.count.col}{Column name which contains the read count per contig <default: 'consensus_count'>}

\item{assay}{what assay type is being used among only two options (bcr,tcr) <default: 'bcr'>}

\item{select.patients}{A vector of select patient ID used to subset the input dataframe <default: NULL>}

\item{expansion.thresh}{Minimum cell count threshold required for a clonotype to be defined as expanded <default: 2>}
}
\value{
A dataframe with node IDs that can be used to create a network.
}
\description{
Takes an input dataframe of VDJ results (from Immcantation pipeline) and creates a node dataframe which can be used to construct an Immune Repertoire network
}
\examples{
node_df <- createRepertoireNodeDataframe(imm_results = bcr_df,expansion.thresh = 2,assay = "bcr")
}
