% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDoublets.R
\name{findDoublets}
\alias{findDoublets}
\title{Annotate Doublets inside a Seurat Object}
\usage{
findDoublets(seurat.obj, sample.col = "sample", threads = 5)
}
\arguments{
\item{seurat.obj}{Seurat Object}

\item{sample.col}{Metadata column name containing sample names within the input Seurat Object <default: 'sample'>}

\item{threads}{number of threads <default: 5>}
}
\value{
A Seurat object with a new column 'doublet_finder' which contains doublet annotations.
}
\description{
Takes an input Seurat Object and annotates all doublets using the DoubletFinder R package
}
\examples{
seurat.obj <- findDoublets(seurat.obj, sample.col = "sample", threads = 10)
}
\references{
McGinnis CS, Murrow LM, Gartner ZJ (2019).
DoubletFinder: Doublet Detection in Single-Cell RNA Sequencing Data Using Artificial Nearest Neighbors.
\emph{Cell Systems} 8, 329–337.
}
