% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCombinedMatrix.R
\name{generateCombinedMatrix}
\alias{generateCombinedMatrix}
\title{Create single cell count matrix}
\usage{
generateCombinedMatrix(
  dataset_loc,
  samples.vec,
  THREADS = 4,
  multi.results = T,
  assay = "gex",
  min.genes.per.cell = 400,
  max.genes.per.cell = NULL
)
}
\arguments{
\item{dataset_loc}{directory path for all cellranger results}

\item{samples.vec}{character vector of all sample names}

\item{THREADS}{number of threads <default: 4>}

\item{multi.results}{bool value for where these are results outputted by cellranger multi <default: TRUE>}

\item{assay}{what assay type is being used among only two options (gex, csp) <default: 'gex'>}

\item{min.genes.per.cell}{minimum number of genes per cell with counts greater than zero <default: 400>}

\item{max.genes.per.cell}{maximum number of genes per cell with counts greater than zero <default: NULL>}
}
\value{
sparse matrix where cells are the columns and genes are the rows.
}
\description{
Takes input directory of cellranger (10x) and combines all samples into one sparse matrix.
This function only works for scRNA-Seq and Cite-Seq data. Each sample sub-directory must contain:
- barcodes.tsv.gz
- features.tsv.gz
- matrix.mtx.gz
}
\examples{
generateCombinedMatrix(dataset_loc, samples.vec,THREADS = 15, multi.results=T,
assay = "gex",min.genes.per.cell = 700,max.genes.per.cell = 2500)
}
