% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGLIPH2inputFiles.R
\name{generateGLIPH2inputFiles}
\alias{generateGLIPH2inputFiles}
\title{Generate input files for GLIPH2}
\usage{
generateGLIPH2inputFiles(
  tcr.df,
  patient.col = "patient",
  clone.col = "clone_id",
  locus.col = "locus",
  unique.id.col = "unique_cell_id",
  vcall.col = "v_call",
  jcall.col = "j_call",
  cdr3.aa.col = "junction_aa",
  results.dir = "./FilesForGLIPH2"
)
}
\arguments{
\item{tcr.df}{TCR VDJ data outputted from the Immcantation pipeline}

\item{patient.col}{Column name which contains the patient ID <default: 'patient'>}

\item{clone.col}{Column name which contains the clone ID <default: 'clone_id'>}

\item{locus.col}{Column name which contains the type of BCR or TCR chain <default: 'locus'>}

\item{unique.id.col}{Column name which contains unique cell IDs <default: 'unique_cell_id'>}

\item{vcall.col}{Column name which contains V gene calls <default: 'v_call'>}

\item{jcall.col}{Column name which contains J gene calls <default: 'j_call'>}

\item{cdr3.aa.col}{Column name which contains the AA sequences of the junction including the CDR3 region <default: 'junction_aa'>}

\item{results.dir}{Path to results directory <default: './FilesForGLIPH2'>}
}
\value{
No variables returned, but a set of files will be outputted to the results directory provided.
}
\description{
Takes an input TCR VDJ data and generates all files needed to run GLIPH2
}
\examples{
generateGLIPH2inputFiles(tcr.df = tcr_df,results.dir = "~/FilesForGLIPH2")
}
\references{
Huang H., Wang C., et al. \emph{Nat Biotechnol} 2020 Apr 27; https://doi.org/10.1038/s41587-020-0505-4
}
