% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateInputFilesImmcantation.R
\name{generateInputFilesImmcantation}
\alias{generateInputFilesImmcantation}
\title{Generate input files for the Immcantation Pipeline}
\usage{
generateInputFilesImmcantation(
  vdj.results.dir,
  contig.annotation.file = "filtered_contig_annotations.csv",
  contig.fasta.file = "filtered_contig.fasta",
  output.file.prefix = "combined",
  contig.id.col = "contig_id",
  output.dir
)
}
\arguments{
\item{vdj.results.dir}{Directory containing VDJ results across all samples}

\item{contig.annotation.file}{String of the file name of the contig annotation results files <default: 'filtered_contig_annotations.csv'>}

\item{contig.fasta.file}{String of the file name of the contig FASTA results files <default: 'filtered_contig.fasta'>}

\item{output.file.prefix}{String of a custom prefix that would be added to the beginning of all files outputted from this function <default: 'combined'>}

\item{contig.id.col}{Column name within the contig annotations containing all contig IDs <default: 'contig_id'>}

\item{output.dir}{Output directory for output files}
}
\value{
A combined output annotation CSV and FASTA file of all VDJ contigs.
}
\description{
Takes an input directory of 10x VDJ results and generates a combined output annotation CSV and FASTA file of all VDJ contigs (BCR or TCR)
}
\examples{
generateInputFilesImmcantation(vdj.results.dir,output.dir,output.file.prefix = "all_BCR")
}
