% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDoublets.R
\name{get10xDoubletRate}
\alias{get10xDoubletRate}
\title{Calculate doublet rate based on sample cell count}
\usage{
get10xDoubletRate(sample_cell_count)
}
\arguments{
\item{sample_cell_count}{Seurat Object}
}
\value{
A sample doublet rate which will get used in the 'findDoublets' function.
}
\description{
Takes and input sample cell count and calculates a doublet rate using published doublet rates from the 3' scRNA-Seq assay (10x Genomics)
}
\examples{
sample.doublet.rate <- get10xDoubletRate(sample_cell_count = 1000)
}
