% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOptimalPCsforSeuratObject.R
\name{plotOptimalPCsforSeuratObject}
\alias{plotOptimalPCsforSeuratObject}
\title{Visualize the optimal PCs for Seurat Object}
\usage{
plotOptimalPCsforSeuratObject(seurat_obj)
}
\arguments{
\item{seurat_obj}{Seurat Object}
}
\value{
A plot showing increasing numbers of PCs where each point represents PCs which cumulatively account for greater than 90% of variation and less than 5% std. dev. of variation between PCs.
}
\description{
Takes an input Seurat Object and generates a plot of variation captured across different numbers of principal components (PC)
}
\examples{
plot <- plotOptimalPCsforSeuratObject(seurat_obj)
}
\references{
This code is from https://hbctraining.github.io/scRNA-seq/lessons/sc_exercises_clustering_analysis.html
}
