% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleAndClusterSeuratObject.R
\name{scaleAndClusterSeuratObject}
\alias{scaleAndClusterSeuratObject}
\title{Scale and Cluster a Seurat Object}
\usage{
scaleAndClusterSeuratObject(
  seurat_obj,
  dims = 1:30,
  npca = 10,
  normalize = TRUE,
  dim.reduction = TRUE,
  tsne = TRUE,
  umap = TRUE
)
}
\arguments{
\item{seurat_obj}{Seurat Object}

\item{dims}{Dimensions of reduction used as input <default: 1:30>}

\item{npca}{Maximum dimensions of reduction for tSNE if number of cells are too small <default: 10>}

\item{normalize}{Bool of whether to normalize using the SCT transform <default: TRUE>}

\item{dim.reduction}{Bool of whether to perform dimension reduction with PCA, tSNE and UMAP <default: TRUE>}

\item{tsne}{Bool of whether to perform dimension reduction with tSNE <default: TRUE>}

\item{umap}{Bool of whether to perform dimension reduction with UMAP <default: TRUE>}
}
\value{
A re-clustered Seurat object.
}
\description{
Takes an input Seurat Object and applies normalization (regression on UMI count) and dimension reduction
}
\examples{
seurat.obj <- scaleAndClusterSeuratObject(seurat.obj,npca = 20,normalize = TRUE,dim.reduction = TRUE)
}
